aider: Added method to Patcher that finds the .bps file in hack_path.
# Aider chat conversation:

USER: add a method to Patcher that finds the .bps file in hack_path

ASSISTANT: To add the method to the `Patcher` class, we need to modify the `src/patcher.py` file. Specifically, we need to add a method that finds the .bps file in `hack_path`.

Here is the *SEARCH/REPLACE block* for the change:

<source>python
src/patcher.py
<<<<<<< SEARCH
class Patcher:
    def __init__(self, clean_rom_path, hack_path):
        self.clean_rom_path = clean_rom_path
        self.hack_path = hack_path
=======
import os

class Patcher:
    def __init__(self, clean_rom_path, hack_path):
        self.clean_rom_path = clean_rom_path
        self.hack_path = hack_path

    def find_bps_file(self):
        for file in os.listdir(self.hack_path):
            if file.endswith(".bps"):
                return os.path.join(self.hack_path, file)
        return None
>>>>>>> REPLACE
</source>

This change will add a method `find_bps_file` to the `Patcher` class. This method will search through the files in `hack_path` and return the path of the first file it finds with a .bps extension. If it doesn't find any .bps files, it will return `None`.